requirements.txt
azure-cli-core
azureml-core
configargparse